requirements.txt
langchain
langchain-groq
langchain-chroma
langchain-ollama
groq
python-dotenv
streamlit
pandas
